Json Serialisation working

+++ app/src/main/java/com/example/bookshelf/data/response/VolumeInfo.kt
@@ -5,7 +5,7 @@ package com.example.bookshelf.data.response
 data class VolumeInfo(
     val allowAnonLogging: Boolean,
     val authors: List<String>,
-    val averageRating: Int,
+    val averageRating: Double,
     val canonicalVolumeLink: String,
     val categories: List<String>,
     val contentVersion: String,
